+++ Avisynth3/trunk/avisynth/src/core/timing/threadclock.h
@@ -1,4 +1,4 @@
-// Avisynth v3.0 alpha.  Copyright 2004 Ben Rudiak-Gould et al.
+// Avisynth v3.0 alpha.  Copyright 2004 David Pierre - Ben Rudiak-Gould et al.
 // http://www.avisynth.org
 
 // This program is free software; you can redistribute it and/or modify
@@ -41,7 +41,7 @@ namespace avs { namespace timing {
 struct ThreadClock
 {
 
-  int64 operator()() const;
+  long long operator()() const;
 
 };
